Corrected Auto Increment to be Int for SQLite | Fix API and test generated API. It works | Begin JWT work

+++ template/create_module_template/mod_xyz/models.py
@@ -8,7 +8,7 @@ from app import db
 class Base(db.Model):
 
     __abstract__ = True
-    id = db.Column(db.BigInteger, autoincrement=True, primary_key=True)
+    id = db.Column(db.Integer, autoincrement=True, primary_key=True)
     created_at = db.Column(db.DateTime, default=db.func.current_timestamp())
     updated_at = db.Column(db.DateTime, default=db.func.current_timestamp(), onupdate=db.func.current_timestamp())
     deleted_at = db.Column(db.DateTime, nullable=True)
